+++ packages/astro-embed-integration/remark-plugin.ts
@@ -1,15 +1,19 @@
 import { type Node, select, selectAll } from 'unist-util-select';
 import blueskyMatcher from '@astro-community/astro-embed-bluesky/matcher';
+import gistMatcher from '@astro-community/astro-embed-gist/matcher';
 import twitterMatcher from '@astro-community/astro-embed-twitter/matcher';
 import vimeoMatcher from '@astro-community/astro-embed-vimeo/matcher';
 import youtubeMatcher from '@astro-community/astro-embed-youtube/matcher';
 import linkPreviewMatcher from '@astro-community/astro-embed-link-preview/matcher';
+import mastodonMatcher from '@astro-community/astro-embed-mastodon/matcher';
 
 const matchers = [
 	[blueskyMatcher, 'BlueskyPost'],
+	[gistMatcher, 'Gist'],
 	[twitterMatcher, 'Tweet'],
 	[vimeoMatcher, 'Vimeo'],
 	[youtubeMatcher, 'YouTube'],
+	[mastodonMatcher, 'MastodonPost'],
 	/** The generic link matcher must be last otherwise it will override other URLs. */
 	[linkPreviewMatcher, 'LinkPreview'],
 ] as const;