+++ packages/astro-embed-mastodon/index.ts
@@ -1,0 +1,1 @@
+export { default as MastodonPost } from './src/MastodonPost.astro';